refactor helpers & testers out of consideration
